On L1 fanout combine UTxOs of same address
This fixes fanning out funds from proxy address to l1 address when the
parties have interacted in the payment channel which resulted in
multiple UTxOs of the same address for which there was a double spent
issue. Not only this fixes the issue, but it is also an optimization to
save on Transactions.